+++ Plugins.properties
@@ -82,15 +82,12 @@ tokenizer=org.omegat.tokenizer.DefaultTokenizer \
     org.omegat.tokenizer.LuceneIndonesianTokenizer \
     org.omegat.tokenizer.LuceneIrishTokenizer \
     org.omegat.tokenizer.LuceneItalianTokenizer \
-    org.omegat.tokenizer.LuceneJapaneseTokenizer \
     org.omegat.tokenizer.LuceneLatvianTokenizer \
     org.omegat.tokenizer.LuceneNorwegianTokenizer \
     org.omegat.tokenizer.LucenePersianTokenizer \
-    org.omegat.tokenizer.LucenePolishTokenizer \
     org.omegat.tokenizer.LucenePortugueseTokenizer \
     org.omegat.tokenizer.LuceneRomanianTokenizer \
     org.omegat.tokenizer.LuceneRussianTokenizer \
-    org.omegat.tokenizer.LuceneSmartChineseTokenizer \
     org.omegat.tokenizer.LuceneSpanishTokenizer \
     org.omegat.tokenizer.LuceneSwedishTokenizer \
     org.omegat.tokenizer.LuceneThaiTokenizer \
