+++ language-modules/zh/build.gradle
@@ -18,6 +18,7 @@ dependencies {
             exclude module: 'jackson-databind'
             exclude group: 'org.jetbrains'
         }
+        implementation(libs.lucene.analyzers.smartcn)
         implementation(libs.languagetool.zh) {
             exclude module: 'languagetool-core'
         }
